Reread the handle list if an entry is bad
- Build successful

desktop_hook.c
- (init_desktop_hook_store)
Add the ability to retry when reading the HANDLEENTRYs.

Prior to this change if two or more handles contained the same HOOK
kernel address (pHead) then this program would exit:

..\desktop_hook.c line 651, init_desktop_hook_store(): Duplicate pHead.

And that is usually because rarely it can occur that a HOOK address is
reused at exactly the moment that this program reads the handle table.
To remedy that this function will now retry reading the handle list for
a full second before exiting, or indefinitely if the force option is
used. For more information review the source comments.

usage.c
- (print_more_options_and_exit)
Update the force option's usage to show that it will also retry reading
the HANDLEENTRY list indefinitely.