Retry thread traversal if it fails
- Build successful

snapshot.c
- (init_snapshot_store)
Change the ability to retry thread traversal.

Prior to this change if NtQuerySystemInformation() failed then this
program would exit unless the force option was used, in which case it
would retry thread traversal every second.

Now if thread traversal fails due to NtQuerySystemInformation() then
this function will by default retry traversal as many times as it can
up to a full second. If a succesful traversal could not be completed by
that time then the program will exit unless the force option is used,
in which case it will retry almost all different types of failed thread
traversals due to NtQuerySystemInformation() indefinitely with no wait.

usage.c
- (print_more_options_and_exit)
Update the force option's usage to show the changed behavior used when
NtQuerySystemInformation() fails.