Ignore duplicate Win32ThreadInfo addresses
- Build successful

snapshot.h
- Add new variable 'unique_w32thread' to the gui struct.
This variable is needed to determine whether or not a GUI thread has a
Win32ThreadInfo kernel address that is unique (not found in any other
GUI thread).

snapshot.c
- (callback_add_gui)
Assume that the Win32ThreadInfo address is unique by default.

- (find_Win32ThreadInfo)
Don't return a GUI thread that has Win32ThreadInfo that is not unique.

- (init_snapshot_store)
Set 'unique_w32thread' to false if a thread has a Win32ThreadInfo
address that's not unique. Prior to this change the program would die
if a duplicate Win32ThreadInfo was encountered.

- (print_gui)
Add print for 'unique_w32thread'.